fixing serialization of patricia tree optime cloneable objects and fix bugs that causes to not encode properly

+++ adrestus-protocol/src/test/java/io/Adrestus/protocol/SyncConsensusExistedTest.java
@@ -36,6 +36,7 @@ import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 
 import java.io.IOException;
+import java.math.BigDecimal;
 import java.math.BigInteger;
 import java.net.InetSocketAddress;
 import java.net.Socket;
@@ -87,7 +88,7 @@ public class SyncConsensusExistedTest {
         ecKeyPair2 = Keys.createEcKeyPair(random);
 
         String address2 = WalletAddress.generate_address((byte) version, ecKeyPair2.getPublicKey());
-        TreeFactory.getMemoryTree(0).store(address2, new PatriciaTreeNode(3000, 0));
+        TreeFactory.getMemoryTree(0).store(address2, new PatriciaTreeNode(BigDecimal.valueOf(3000), 0));
         Thread.sleep(2000);
         ECDSASignatureData signatureData2 = ecdsaSign.secp256SignMessage(HashUtil.sha256(StringUtils.getBytesUtf8(address2)), ecKeyPair2);
         kad2 = new KademliaData(new SecurityAuditProofs(address2, vk2, ecKeyPair2.getPublicKey(), signatureData2), new NettyConnectionInfo("192.168.1.117", KademliaConfiguration.PORT));
